added weak self to all api/parse requests and cleaned up the code (spacing/comments)

+++ favorites/View Controllers/Profile/ProfileViewController.m
@@ -64,19 +64,21 @@ static NSString *segueToLikes = @"showLikes";
     [query whereKey:@"author" equalTo:self.user];
     [query orderByDescending:@"createdAt"];
     query.limit = 20;
-
-    // fetch data asynchronously
+    typeof(self) __weak weakSelf = self;
     [query findObjectsInBackgroundWithBlock:^(NSArray *updates, NSError *error) {
-        if (updates != nil) {
-            self.updates = updates;
-            NSLog(@"got updates");
-            self.isLikedByUser = [[NSMutableDictionary alloc] init];
-            for (Update *update in self.updates) {
-                [self.isLikedByUser setValue:@"0" forKey:update.objectId];
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (updates != nil) {
+                self.updates = updates;
+                NSLog(@"got updates");
+                self.isLikedByUser = [[NSMutableDictionary alloc] init];
+                for (Update *update in self.updates) {
+                    [self.isLikedByUser setValue:@"0" forKey:update.objectId];
+                }
+                [self getLikes];
+            } else {
+                NSLog(@"%@", error.localizedDescription);
             }
-            [self getLikes];
-        } else {
-            NSLog(@"%@", error.localizedDescription);
         }
     }];
 }
@@ -90,19 +92,22 @@ static NSString *segueToLikes = @"showLikes";
     PFQuery *query = [PFQuery orQueryWithSubqueries:@[queryUser1,queryUser2]];
     NSArray *keys = @[@"user1", @"user2"];
     [query includeKeys:keys];
-
+    typeof(self) __weak weakSelf = self;
     [query findObjectsInBackgroundWithBlock:^(NSArray *friends, NSError *error) {
-        if (friends != nil) {
-            self.friends = [[NSMutableArray alloc] init];
-            NSLog(@"# of friends: %lu", (unsigned long)[friends count]);
-            for (Friend *friend in friends) {
-                if ([friend.user1.objectId isEqual:self.user.objectId]) {
-                    [self.friends addObject:friend.user2];
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (friends != nil) {
+                strongSelf.friends = [[NSMutableArray alloc] init];
+                NSLog(@"# of friends: %lu", (unsigned long)[friends count]);
+                for (Friend *friend in friends) {
+                    if ([friend.user1.objectId isEqual:strongSelf.user.objectId]) {
+                        [strongSelf.friends addObject:friend.user2];
+                    }
+                    else [strongSelf.friends addObject:friend.user1];
                 }
-                else [self.friends addObject:friend.user1];
+            } else {
+                NSLog(@"%@", error.localizedDescription);
             }
-        } else {
-            NSLog(@"%@", error.localizedDescription);
         }
     }];
 }
@@ -113,16 +118,20 @@ static NSString *segueToLikes = @"showLikes";
     [query includeKeys:keys];
     [query whereKey:@"user" equalTo:[PFUser currentUser]];
     query.limit = 20;
+    typeof(self) __weak weakSelf = self;
     [query findObjectsInBackgroundWithBlock:^(NSArray *likes, NSError *error) {
-        if (likes != nil) {
-            NSLog(@"got likes");
-            for (Like *like in likes) {
-                [self.isLikedByUser setValue:@"1" forKey:like.update.objectId];
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (likes != nil) {
+                NSLog(@"got likes");
+                for (Like *like in likes) {
+                    [strongSelf.isLikedByUser setValue:@"1" forKey:like.update.objectId];
+                }
+                [strongSelf.tableView reloadData];
+                [SVProgressHUD dismiss];
+            } else {
+                NSLog(@"%@", error.localizedDescription);
             }
-            [self.tableView reloadData];
-            [SVProgressHUD dismiss];
-        } else {
-            NSLog(@"%@", error.localizedDescription);
         }
     }];
 }
@@ -178,50 +187,61 @@ static NSString *segueToLikes = @"showLikes";
     [self presentViewController:editUpdate animated:YES completion:nil];
 }
 - (void) deleteUpdate:(Update *)update {
+    typeof(self) __weak weakSelf = self;
     [update deleteInBackgroundWithBlock:^(BOOL succeeded, NSError * _Nullable error) {
-        if (succeeded) {
-            NSLog(@"deleted post %@", update.caption);
-            // delete all likes and comments
-            [self deleteLikes:update];
-        }
-        else {
-            NSLog(@"problem deleting post: %@", error.localizedDescription);
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (succeeded) {
+                NSLog(@"deleted post %@", update.caption);
+                [self deleteLikes:update]; // delete all likes and comments
+            }
+            else {
+                NSLog(@"problem deleting post: %@", error.localizedDescription);
+            }
         }
     }];
 }
 -(void) deleteLikes:(Update *)update {
     PFQuery *query = [PFQuery queryWithClassName:@"Like"];
     [query whereKey:@"update" equalTo:update];
+    typeof(self) __weak weakSelf = self;
     [query findObjectsInBackgroundWithBlock:^(NSArray *likes, NSError *error) {
-        if (likes != nil) {
-            [Like deleteAllInBackground:likes block:^(BOOL succeeded, NSError * _Nullable error) {
-                if (succeeded) {
-                    NSLog(@"deleted likes");
-                    [self deleteComments:update];
-                } else {
-                    NSLog(@"%@", error.localizedDescription);
-                }
-            }];
-        } else {
-            NSLog(@"%@", error.localizedDescription);
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (likes != nil) {
+                [Like deleteAllInBackground:likes block:^(BOOL succeeded, NSError * _Nullable error) {
+                    if (succeeded) {
+                        NSLog(@"deleted likes");
+                        [strongSelf deleteComments:update];
+                    } else {
+                        NSLog(@"%@", error.localizedDescription);
+                    }
+                }];
+            } else {
+                NSLog(@"%@", error.localizedDescription);
+            }
         }
     }];
 }
 -(void) deleteComments:(Update *)update {
     PFQuery *query = [PFQuery queryWithClassName:@"Comment"];
     [query whereKey:@"update" equalTo:update];
+    typeof(self) __weak weakSelf = self;
     [query findObjectsInBackgroundWithBlock:^(NSArray *comments, NSError *error) {
-        if (comments != nil) {
-            [Comment deleteAllInBackground:comments block:^(BOOL succeeded, NSError * _Nullable error) {
-                if (succeeded) {
-                    NSLog(@"deleted comments");
-                    [self getUpdates];
-                } else {
-                    NSLog(@"%@", error.localizedDescription);
-                }
-            }];
-        } else {
-            NSLog(@"%@", error.localizedDescription);
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (comments != nil) {
+                [Comment deleteAllInBackground:comments block:^(BOOL succeeded, NSError * _Nullable error) {
+                    if (succeeded) {
+                        NSLog(@"deleted comments");
+                        [strongSelf getUpdates];
+                    } else {
+                        NSLog(@"%@", error.localizedDescription);
+                    }
+                }];
+            } else {
+                NSLog(@"%@", error.localizedDescription);
+            }
         }
     }];
 }
@@ -238,13 +258,17 @@ static NSString *segueToLikes = @"showLikes";
             }
         }];
         [self.isLikedByUser setValue:@"1" forKey:update.objectId];
+        typeof(self) __weak weakSelf = self;
         [Like createLike:[PFUser currentUser] onUpdate:update withCompletion:^(BOOL succeeded, NSError * _Nullable error) {
-            if (succeeded) {
-                NSLog(@"created new like!");
-                [self sendLikePush:update];
-                [self getUpdates];
-            } else {
-                NSLog(@"%@", error.localizedDescription);
+            typeof(weakSelf) strongSelf = weakSelf;
+            if (strongSelf) {
+                if (succeeded) {
+                    NSLog(@"created new like!");
+                    [strongSelf sendLikePush:update];
+                    [strongSelf getUpdates];
+                } else {
+                    NSLog(@"%@", error.localizedDescription);
+                }
             }
         }];
     }
@@ -266,19 +290,23 @@ static NSString *segueToLikes = @"showLikes";
     PFQuery *query = [PFQuery queryWithClassName:@"Like"];
     [query whereKey:@"user" equalTo:[PFUser currentUser]];
     [query whereKey:@"update" equalTo:update];
+    typeof(self) __weak weakSelf = self;
     [query findObjectsInBackgroundWithBlock:^(NSArray *likes, NSError *error) {
-        if (likes != nil) {
-            Like *like = likes[0];
-            [like deleteInBackgroundWithBlock:^(BOOL succeeded, NSError * _Nullable error) {
-                if (succeeded) {
-                    NSLog(@"deleted like");
-                    [self getUpdates];
-                } else {
-                    NSLog(@"%@", error.localizedDescription);
-                }
-            }];
-        } else {
-            NSLog(@"%@", error.localizedDescription);
+        typeof(weakSelf) strongSelf = weakSelf;
+        if (strongSelf) {
+            if (likes != nil) {
+                Like *like = likes[0];
+                [like deleteInBackgroundWithBlock:^(BOOL succeeded, NSError * _Nullable error) {
+                    if (succeeded) {
+                        NSLog(@"deleted like");
+                        [strongSelf getUpdates];
+                    } else {
+                        NSLog(@"%@", error.localizedDescription);
+                    }
+                }];
+            } else {
+                NSLog(@"%@", error.localizedDescription);
+            }
         }
     }];
 }
